chore: reorganize some modules
