+++ pkg/constant.go
@@ -1,4 +1,4 @@
-package plutus
+package pkg
 
 import (
 	"math/big"
